services/worker-service/requirements.txt
psycopg2-binary
boto3
redis
pymupdf
python-dotenv